일단 완료

+++ src/main/java/com/yein/rest/TodoServiceImpl.java
@@ -3,11 +3,7 @@ package com.yein.rest;
 import java.util.List;
 
 import javax.ws.rs.Consumes;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.GET;
-import javax.ws.rs.NotFoundException;
 import javax.ws.rs.POST;
-import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
@@ -26,15 +22,12 @@ public class TodoServiceImpl implements TodoService {
 	private TodoMapper mapper;
 
 	// 목록조회
-	@GET
+	@POST
 	@Path("/list")
 	@Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
 	public Response getList() {
 		List<Todo> todos = mapper.getList();
-		System.out.println(todos.toString());
-		return Response.status(200).
-				header("Content-Type", "application/json;charset=UTF-8")
-				.entity(todos + " list success").build();
+		return Response.status(200).header("Content-Type", "application/json;charset=UTF-8").entity(todos).build();
 	}
 
 	// 삽입
@@ -43,44 +36,60 @@ public class TodoServiceImpl implements TodoService {
 	@Consumes(MediaType.APPLICATION_JSON + ";charset=utf-8")
 	@Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
 	public Response insert(Todo todo) {
-		mapper.insert(todo);
-		return Response.status(200)
-				.header("Content-Type", "application/json;charset=UTF-8")
-				.entity(todo.toString() + " insert success").build();
+		if (mapper.read(todo) != null) {
+			return Response.status(200).header("Content-Type", "application/json;charset=UTF-8")
+					.entity(todo.toString() + " \n 삽입 실패. 데이터가 이미 존재합니다.").build();
+		} else if (todo.getRegDate().after(todo.getExpDate())) {
+			return Response.status(200).header("Content-Type", "application/json;charset=UTF-8")
+					.entity(todo.toString() + " \n 삽입 실패. 만료일은 등록일 이후여야 합니다.").build();
+		} else {
+			mapper.insert(todo);
+			return Response.status(200).header("Content-Type", "application/json;charset=UTF-8")
+					.entity(todo.toString() + " \n 삽입 성공").build();
+		}
 	}
 
 	// 조회
-	@GET
+	@POST
 	@Path("/read")
 	@Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
-	public Response read(Todo todo) throws NotFoundException {
-		mapper.read(todo);
-		return Response.status(200)
-				.header("Content-Type", "application/json;charset=UTF-8")
-				.entity(todo).build();
+	public Response read(Todo todo) {
+		if ( mapper.read(todo) == null) {
+			return Response.status(200).header("Content-Type", "application/json;charset=UTF-8")
+					.entity(todo.getNo() +"번 데이터  조회 실패. \n데이터가 없습니다.").build();
+		} else
+			return Response.status(200).header("Content-Type", "application/json;charset=UTF-8").entity(todo).build();
 	}
 
 	// 수정
-	@PUT
+	@POST
 	@Path("/update")
 	@Consumes(MediaType.APPLICATION_JSON + ";charset=utf-8")
 	@Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
 	public Response update(Todo todo) {
-		mapper.update(todo);
-		return Response.status(200)
-				.header("Content-Type", "application/json;charset=UTF-8")
-				.entity(todo.toString() + " update success").build();
+		if (mapper.read(todo) == null) {
+			return Response.status(200).header("Content-Type", "application/json;charset=UTF-8")
+					.entity(todo.toString() + " \n 수정 실패. 데이터가 없습니다.").build();
+		} else {
+			mapper.update(todo);
+			return Response.status(200).header("Content-Type", "application/json;charset=UTF-8")
+					.entity(todo.toString() + " 수정 성공").build();
+		}
 	}
 
 	// 삭제
-	@DELETE
+	@POST
 	@Path("/delete")
 	@Consumes(MediaType.APPLICATION_JSON + ";charset=utf-8")
 	@Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
 	public Response delete(Todo todo) {
-		mapper.delete(todo);
-		return Response.status(200)
-				.header("Content-Type", "application/json;charset=UTF-8")
-				.entity(todo.toString() + " delete success").build();
+		if (mapper.read(todo) == null) {
+			return Response.status(200).header("Content-Type", "application/json;charset=UTF-8")
+					.entity(todo.getNo() +"번 데이터  삭제 실패.\n 데이터가 없습니다.").build();
+		} else {
+			mapper.delete(todo);
+			return Response.status(200).header("Content-Type", "application/json;charset=UTF-8")
+					.entity(todo.getNo() +"번 데이터  삭제 성공").build();
+		}
 	}
 }
